# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document APIs",
        "contact": {}
    },
    "paths": {
...
        "/cms/v1/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Roles",
                "parameters": [
                    {
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "DataStatusEnable",
                            "DataStatusDisable"
                        ],
                        "name": "data_status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at.desc",
                        "name": "sorts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.RoleCmsDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Create Role",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.RoleCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.RoleCmsDto"
                        }
                    }
                }
            }
        },
        "/cms/v1/roles/{role_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Update Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "role_id",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.RoleCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.RoleCmsDto"
                        }
                    }
                }
            }
        },
...
        "/storage/v1/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload Image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/storage/v1/images/{filename}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Download Image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/storage/v1/videos": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload Video",
                "parameters": [
                    {
                        "type": "file",
                        "description": "files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/storage/v1/videos/{filename}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Download Video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "db.RoleCmsData": {
            "type": "object",
            "required": [
                "data_status",
                "name",
                "permissions"
            ],
            "properties": {
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "permissions": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/enum.Permission"
                    }
                }
            }
        },
        "db.RoleCmsDto": {
            "type": "object",
            "properties": {
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.Permission"
                    }
                },
                "role_id": {
                    "type": "string",
                    "example": "671db9eca1f1b1bdbf3d4617"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T03:04:05Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "editor"
                }
            }
        },
...
        "enum.DataStatus": {
            "type": "string",
            "enum": [
                "enable",
                "disable"
            ],
            "x-enum-varnames": [
                "DataStatusEnable",
                "DataStatusDisable"
            ]
        },
        "enum.Permission": {
            "type": "string",
            "enum": [
                "system_setting",
                "system_audit_log",
                "client_view",
                "client_create",
                "client_delete",
                "role_view",
                "role_create",
                "role_update",
                "user_view",
                "user_create",
                "user_update",
                "tenant_view",
                "tenant_create",
                "tenant_update"
            ],
            "x-enum-varnames": [
                "PermissionSystemSetting",
                "PermissionSystemAuditLog",
                "PermissionClientView",
                "PermissionClientCreate",
                "PermissionClientDelete",
                "PermissionRoleView",
                "PermissionRoleCreate",
                "PermissionRoleUpdate",
                "PermissionUserView",
                "PermissionUserCreate",
                "PermissionUserUpdate",
                "PermissionTenantView",
                "PermissionTenantCreate",
                "PermissionTenantUpdate"
            ]
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Header"
        }
    }
}